pyproject.toml
[tool.poetry]
name = "coursework-drf"
version = "0.1.0"
description = ""
authors = ["kifusss <kifus_06@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
drf-yasg = "^1.21.7"
setuptools = "^70.0.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
